refactor: 리뷰 페이지 상태 관리와 섹션 UI 분리
