+++ apps/web/src/views/review/ui/ReviewEnvironmentSection.tsx
@@ -1,0 +1,75 @@
+import { Chip, Field, Icon } from '@plog/ui';
+import { cn } from '@plog/utils';
+
+import { REVIEW_ENVIRONMENT_GROUPS } from '../model/constants';
+import { type ReviewEnvironmentGroup } from '../model/types';
+import { type ReviewFormController } from '../model/use-review-page';
+
+type ReviewEnvironmentSectionProps = {
+  controller: ReviewFormController;
+};
+
+function EnvironmentChoiceGroup({
+  group,
+  value,
+  onChange,
+}: {
+  group: ReviewEnvironmentGroup;
+  value: string | null;
+  onChange: (value: string | null) => void;
+}) {
+  return (
+    <div className="flex w-full flex-col gap-3">
+      <div className="body-lg flex items-center gap-1.5 text-semantic-object-bold">
+        <Icon
+          name={group.iconName}
+          size={20}
+          className="text-semantic-object-subtle"
+        />
+        {group.title}
+      </div>
+      <input type="hidden" name={group.name} value={value ?? ''} />
+      <div className="grid grid-cols-3 gap-3">
+        {group.options.map((option) => (
+          <Chip
+            key={option}
+            size="large"
+            variant="solid"
+            pressed={value === option}
+            className="w-full min-w-0"
+            onPressedChange={(pressed) => onChange(pressed ? option : null)}
+          >
+            <span className={cn('label-sm')}>{option}</span>
+          </Chip>
+        ))}
+      </div>
+    </div>
+  );
+}
+
+export default function ReviewEnvironmentSection({
+  controller,
+}: ReviewEnvironmentSectionProps) {
+  const { environmentValues, handleEnvironmentChange } = controller;
+
+  return (
+    <section className="flex flex-col px-6 pt-6 pb-10">
+      <Field
+        label="방문하신 장소의 환경은 어떠셨나요?"
+        className="gap-4"
+        required
+      >
+        <div className="flex flex-col gap-6">
+          {REVIEW_ENVIRONMENT_GROUPS.map((group) => (
+            <EnvironmentChoiceGroup
+              key={group.name}
+              group={group}
+              value={environmentValues[group.name]}
+              onChange={(value) => handleEnvironmentChange(group.name, value)}
+            />
+          ))}
+        </div>
+      </Field>
+    </section>
+  );
+}
