+++ apps/web/src/views/review/ui/SectionDivider.tsx
@@ -1,0 +1,3 @@
+export default function SectionDivider() {
+  return <div aria-hidden="true" className="h-2 bg-semantic-bg-deep" />;
+}
